requirements.txt
# Bitcoin Market Analysis Tool Dependencies

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
ta>=0.10.2  # Technical Analysis library
pandas-ta>=0.3.14b  # Additional TA indicators

# Binance API
python-binance>=1.0.19
websockets>=12.0

# Audio notifications (handled by browser for web interface)
# CLI mode uses macOS 'afplay' or terminal bell - no extra dependency needed

# Configuration and utilities
python-dotenv>=1.0.0
rich>=13.0.0  # Beautiful terminal output

# Optional: For advanced pattern recognition
scipy>=1.11.0

# Web interface
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.35.0
